fix: question shared between a room is now the same
